Restucture pallet airdrop
