+++ pallets/airdrop/src/lib.rs
@@ -1,52 +1,3 @@
-//! # Airdrop pallet
-//! This pallet is responsible to do airdropping to claimers of icon_address.
-//! - [`Config`]
-//! - [`Call`]
-//!
-//! ## Interface
-//!
-//! ### Disptachable function for end-user
-//!
-//! * [`claim_request`]
-//!
-//! Signed call that accepts set of ice data (icon signature, initial message, icon address).
-//! This maps the ice address with icon address ( if previously not done)
-//! and register this request in pending claims.
-//!
-//! * [`cancel_request`]
-//!
-//! Allows any user to cancel last claim request. This does so by just removing
-//! respective entry from pending queue. Only callable by the owner who
-//! first called `claim_request` or root account or palelt authorised account.
-//!
-//! * [`donateto_creditor`]
-//!
-//! Allows any user to donate funds to the pallet account. Pallet account is
-//! the account from which the airdrop funds will be transferred from.
-//! This dispatchable should be called to credit pallet with enough balance
-//! else every call to `complete_transfer` will fail. This should be done before any
-//! user arrives to node for claiming request
-//!
-//! ---
-//! ### Dispatchable calls for internal use
-//! * [`complete_transfer`]
-//!
-//! This transfer the fund to given ice address with given transaction details
-//! inside `ServerResponse` type. If transferring the fund succeed, it will also
-//! remove the queue from pendingClaims and update any snapshot info as needed.
-//! This is only callable by sudo/root account
-//!
-//! * [`register_failed_claim`]
-//! This dispatchable provides a mean to register that some entry was failed to process
-//! and this responsibility is to be taken by node operator.
-//!
-//! * [`update_processed_upto_counter`]
-//! To update the mark that records upto which block of claiming has been completed
-//!
-//! * [`remove_from_pending_claim`]
-//! To remove something from pending claim as it is no longer needed. For eg: when record
-//! claiming account of this icon address is 0 token
-
 #![cfg_attr(not(feature = "std"), no_std)]
 pub use pallet::*;
 #[cfg(test)]
@@ -58,8 +9,11 @@ pub mod tests;
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
 
-/// All the types and alises must be defined here
+/// All the types, traits defination and alises are inside this
 pub mod types;
+
+/// All independent utilities function are inside here
+pub mod utils;
 
 /// An identifier for a type of cryptographic key.
 /// For this pallet, account associated with this key must be same as
@@ -79,7 +33,7 @@ pub const DEFAULT_RETRY_COUNT: u8 = 2;
 
 #[frame_support::pallet]
 pub mod pallet {
-	use super::types;
+	use super::{types, utils};
 	use sp_runtime::traits::Saturating;
 
 	use frame_support::pallet_prelude::*;
@@ -90,9 +44,6 @@ pub mod pallet {
 		Currency, ExistenceRequirement, Hooks, LockableCurrency, ReservableCurrency,
 	};
 	use frame_system::offchain::CreateSignedTransaction;
-	use sp_runtime::traits::{
-		AtLeast32BitUnsigned, CheckedAdd, CheckedDiv, CheckedMul, CheckedSub, Convert,
-	};
 	use types::IconVerifiable;
 
 	/// Configure the pallet by specifying the parameters and types on which it depends.
@@ -795,57 +746,13 @@ pub mod pallet {
 				return Err(ClaimError::HttpError);
 			}
 
-			// first try to deserialize into expected ok struct
-			// then try to deserialize into known error
-			// else error an invalid format
-			let deserialize_response_res =
-				serde_json::from_slice::<types::ServerResponse>(response_bytes.as_slice());
-
-			match deserialize_response_res {
-				Ok(response) => Ok(response),
-
-				Err(_) => {
-					let deserialize_error_res =
-						serde_json::from_slice::<types::ServerError>(response_bytes.as_slice());
-
-					match deserialize_error_res {
-						Ok(server_error) => Err(ClaimError::ServerError(server_error)),
-						Err(_) => Err(ClaimError::InvalidResponse),
-					}
-				}
-			}
+			utils::unpack_server_response(response_bytes.as_slice())
 		}
 
 		/// Return an indicater (bool) on weather the offchain worker
 		/// should be run on this block number or not
 		pub fn should_run_on_this_block(block_number: types::BlockNumberOf<T>) -> bool {
 			block_number % crate::OFFCHAIN_WORKER_BLOCK_GAP.into() == 0_u32.into()
-		}
-
-		/// Returns the amount to be transferred from given serverResponse
-		pub fn get_total_amount(
-			server_response: &types::ServerResponse,
-		) -> Result<types::BalanceOf<T>, &'static str> {
-			use sp_runtime::traits::CheckedAdd;
-
-			let amount: types::BalanceOf<T> = server_response
-				.amount
-				.try_into()
-				.map_err(|_| "Failed to convert server_response.amount to balance type")?;
-			let stake: types::BalanceOf<T> = server_response
-				.stake
-				.try_into()
-				.map_err(|_| "Failed to convert server_response.stake to balance type")?;
-			let omm: types::BalanceOf<T> = server_response
-				.omm
-				.try_into()
-				.map_err(|_| "Failed to convert server_response.oom to balance type")?;
-
-			amount
-				.checked_add(&stake)
-				.ok_or("Adding server_response(amount+stake) overflowed")?
-				.checked_add(&omm)
-				.ok_or("Adding server_response(amount+stake+omm) overflowed")
 		}
 	}
 
@@ -969,12 +876,13 @@ pub mod pallet {
 			// Block number after which enable to do vesting
 			const VESTING_APPLICABLE_FROM: u32 = 100u32;
 
-			let total_amount = server_response
-				.amount
-				.checked_add(server_response.stake)
-				.ok_or(sp_runtime::ArithmeticError::Overflow)?
-				.checked_add(server_response.omm)
-				.ok_or(sp_runtime::ArithmeticError::Overflow)?;
+			let total_amount = utils::get_response_sum(&server_response).map_err(|e| {
+				log::error!(
+					"[Airdrop pallet] Cannot get total sum from server response. {:?}",
+					e
+				);
+				e
+			})?;
 
 			let (instant_amount, vesting_amount): (
 				types::BalanceOf<T>,
@@ -1000,7 +908,7 @@ pub mod pallet {
 			})?;
 
 			// Get primary & secondary vesting to be applied
-			let [primary_vesting, secondary_vesting] = types::new_vesting_with_deadline::<
+			let [primary_vesting, secondary_vesting] = utils::new_vesting_with_deadline::<
 				T,
 				VESTING_APPLICABLE_FROM,
 			>(vesting_amount, BLOCKS_IN_YEAR.into());
@@ -1081,114 +989,5 @@ pub mod airdrop_crypto {
 		type RuntimeAppPublic = Public;
 		type GenericSignature = sp_core::sr25519::Signature;
 		type GenericPublic = sp_core::sr25519::Public;
-	}
-}
-
-/// Implement IconVerifiable for Anything that can be decoded into Vec<u8>
-// However not that
-impl types::IconVerifiable for sp_runtime::AccountId32 {
-	/// Function to make sure that icon_address, ice_address and message are in sync
-	/// On a high level, it does so by checking for these two verification
-	/// * Make sure that icon_signature is equal to or greater than 65
-	/// * Make sure that the ice_address encoded in the message and passed
-	///    in this function (i.e dispatchable from where this function is called)
-	///    are same
-	/// * Make sure that this message is signed by the same icon_address
-	///    that is being passed to this function (i.e caller for this function)
-	///
-	/// @return:
-	/// verbose error type which point exactly where the process failed
-	///
-	/// @parameter:
-	/// * ice_address: ss58 encoded bytes of origin of parent dispatchable
-	/// * icon_address: icon_address
-	/// * icon_signature: icon signature
-	/// * message: raw message
-	fn verify_with_icon(
-		&self,
-		icon_wallet: &types::IconAddress,
-		icon_signature: &[u8],
-		message: &[u8],
-	) -> Result<(), types::SignatureValidationError> {
-		use codec::Encode;
-		use fp_evm::LinearCostPrecompile;
-		use frame_support::ensure;
-		use pallet_evm_precompile_sha3fips::Sha3FIPS256;
-		use pallet_evm_precompile_simple::ECRecoverPublicKey;
-		use types::SignatureValidationError;
-
-		const COST: u64 = 1;
-		const PADDING_FOR_V: [u8; 31] = [0; 31];
-
-		let ice_address = hex::encode(self.encode());
-		let ice_address = ice_address.as_bytes();
-
-		/* =======================================
-				Validate the icon_signature length
-		*/
-		ensure!(
-			icon_signature.len() == 65,
-			SignatureValidationError::InvalidIconSignature
-		);
-		// === verified the length of icon_signature
-
-		/* ======================================================
-			Verify that the message constains the same ice_address
-			as being passed to this function
-		*/
-		let extracted_ice_address = {
-			// TODO:
-			// make sure that message will always be in expected format
-			const PREFIX_LEN: usize =
-				b"ice_sendTransaction.data.{method.transfer.params.{wallet.".len();
-			let address_len = ice_address.len();
-			&message[PREFIX_LEN..PREFIX_LEN + address_len]
-		};
-
-		ensure!(
-			&ice_address == &extracted_ice_address,
-			SignatureValidationError::InvalidIceAddress
-		);
-		// ==== Verfiied that ice_address in encoded message
-		// and recived in function parameterare same
-
-		/* ================================================
-			verify thet this message is being signed by same
-			icon_address as passed in this function
-		*/
-		let (_exit_status, message_hash) = Sha3FIPS256::execute(&message, COST)
-			.map_err(|_| SignatureValidationError::Sha3Execution)?;
-		let formatted_icon_signature = {
-			let sig_r = &icon_signature[..32];
-			let sig_s = &icon_signature[32..64];
-			let sig_v = &icon_signature[64..];
-
-			// Sig final is in the format of:
-			// object hash + 31 byte padding + 1 byte v + 32 byte r + 32 byte s
-			message_hash
-				.iter()
-				.chain(&PADDING_FOR_V)
-				.chain(sig_v)
-				.chain(sig_r)
-				.chain(sig_s)
-				.cloned()
-				.collect::<sp_std::vec::Vec<u8>>()
-		};
-
-		let (_exit_status, icon_pub_key) =
-			ECRecoverPublicKey::execute(&formatted_icon_signature, COST)
-				.map_err(|_| SignatureValidationError::InvalidIconSignature)?;
-
-		let (_exit_status, computed_icon_address) = Sha3FIPS256::execute(&icon_pub_key, COST)
-			.map_err(|_| SignatureValidationError::Sha3Execution)?;
-
-		ensure!(
-			&computed_icon_address[computed_icon_address.len() - 20..] == icon_wallet.as_slice(),
-			SignatureValidationError::InvalidIconAddress
-		);
-		// ===== It is now verified that the message is signed by same icon address
-		// as passed in this function
-
-		Ok(())
 	}
 }
