+++ pallets/airdrop/src/tests/mod.rs
@@ -16,7 +16,7 @@ pub mod prelude {
 	};
 	pub use hex_literal::hex as decode_hex;
 	pub use pallet_airdrop::mock::{self, AirdropModule, Origin, Test};
-	pub use pallet_airdrop::types;
+	pub use pallet_airdrop::{types, utils};
 	pub use sp_core::bytes;
 
 	pub use sp_runtime::traits::{Bounded, IdentifyAccount, Saturating};
